payment-recovery/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
twilio==9.0.5
slack_sdk==3.27.1
python-multipart==0.0.9

# AI message generation (optional — falls back to templates if key not set)
openai>=1.0.0

# MCP-based Slack alerting (slack_mcp.py)
anthropic>=0.84.0
mcp>=1.0.0
